[features - 21] - Add GetDropdowns for driver method and update data types

+++ DriveSync/DriveSync/Constants/CommonConstants.cs
@@ -2,12 +2,12 @@
 {
     public class CommonConstants
     {
-        public const int ASSIGNMENT_ASSIGNED = 1;
-        public const int ASSIGNMENT_RETURNED = 2;
+        public const short ASSIGNMENT_ASSIGNED = 1;
+        public const short ASSIGNMENT_RETURNED = 2;
 
 
-        public const int VEHICLE_ASSIGNED = 2;
+        public const short VEHICLE_ASSIGNED = 2;
 
-        public const int DRIVER_ASSIGNED = 2;
+        public const short DRIVER_ASSIGNED = 2;
     }
 }
